refactor : InputView 클래스를 통한 입력 로직 분리
